requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
jinja2
python-multipart